add oceanitems file
