+++ Assets/Scripts/Game/Defined/Serialized/Items/Equippable.cs
@@ -29,17 +29,4 @@ namespace Scripts.Game.Defined.Serialized.Items {
             return new DamageResist();
         }
     }
-
-    public class FishHook : EquippableItem {
-
-        public FishHook() : base(EquipType.WEAPON, 50, "Fish Hook", "A used fish hook.") {
-            AddFlatStatBonus(StatType.STRENGTH, 5);
-            AddFlatStatBonus(StatType.AGILITY, 1);
-            AddFlatStatBonus(StatType.VITALITY, -1);
-        }
-
-        public override Buff CreateBuff() {
-            return new FishShook();
-        }
-    }
 }
