+++ Assets/Scripts/Game/Defined/Serialized/Items/OceanItems.cs
@@ -1,0 +1,20 @@
+﻿using Scripts.Game.Defined.Serialized.Buffs;
+using Scripts.Model.Buffs;
+using Scripts.Model.Items;
+using Scripts.Model.Stats;
+
+namespace Scripts.Game.Defined.Serialized.Items {
+
+    public class FishHook : EquippableItem {
+
+        public FishHook() : base(EquipType.WEAPON, 50, "Fish Hook", "A used fish hook.") {
+            AddFlatStatBonus(StatType.STRENGTH, 5);
+            AddFlatStatBonus(StatType.AGILITY, 1);
+            AddFlatStatBonus(StatType.VITALITY, -1);
+        }
+
+        public override Buff CreateBuff() {
+            return new FishShook();
+        }
+    }
+}
