+++ Assets/Scripts/Game/Defined/Serialized/Items/OceanItems.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: fe1f6189ff74a6246aec17920760177d
+timeCreated: 1507774912
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
